sources: refactor base sources into module.
Extract Base and FileBase sources into `sources/_base.py`.

LP: #1648906
