+++ snapcraft/tests/sources/test_base.py
@@ -1,0 +1,100 @@
+# -*- Mode:Python; indent-tabs-mode:nil; tab-width:4 -*-
+#
+# Copyright (C) 2016 Canonical Ltd
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+import os
+import unittest.mock
+
+from snapcraft.internal import sources
+from snapcraft import tests
+
+
+class TestFileBase(tests.TestCase):
+
+    def get_mock_file_base(self, source, dir):
+        file_src = sources.FileBase(source, dir)
+        setattr(file_src, "provision", unittest.mock.Mock())
+        return file_src
+
+    @unittest.mock.patch('snapcraft.internal.sources.FileBase.download')
+    def test_pull_url(self, mock_download):
+        file_src = self.get_mock_file_base(
+            'http://snapcraft.io/snapcraft.yaml', 'dir')
+        file_src.pull()
+
+        mock_download.assert_called_once_with()
+        file_src.provision.assert_called_once_with(file_src.source_dir)
+
+    @unittest.mock.patch('shutil.copy2')
+    def test_pull_copy(self, mock_shutil_copy2):
+        file_src = self.get_mock_file_base('snapcraft.yaml', 'dir')
+        file_src.pull()
+
+        mock_shutil_copy2.assert_called_once_with(
+            file_src.source, file_src.source_dir)
+        file_src.provision.assert_called_once_with(file_src.source_dir)
+
+    @unittest.mock.patch('snapcraft.internal.sources._base.requests')
+    @unittest.mock.patch(
+        'snapcraft.internal.sources._base.download_requests_stream')
+    @unittest.mock.patch(
+        'snapcraft.internal.sources._base.download_urllib_source')
+    def test_download_file_destination(self, dus, drs, req):
+        file_src = self.get_mock_file_base(
+            'http://snapcraft.io/snapcraft.yaml', 'dir')
+        self.assertFalse(hasattr(file_src, "file"))
+
+        file_src.pull()
+
+        self.assertEqual(file_src.file, os.path.join(
+                file_src.source_dir, os.path.basename(file_src.source)))
+
+    @unittest.mock.patch(
+        'snapcraft.internal.sources._base.download_requests_stream')
+    @unittest.mock.patch('snapcraft.internal.sources._base.requests')
+    def test_download_http(self, mock_requests, mock_download):
+        file_src = self.get_mock_file_base(
+            'http://snapcraft.io/snapcraft.yaml', 'dir')
+
+        mock_request = unittest.mock.Mock()
+        mock_requests.get.return_value = mock_request
+
+        file_src.pull()
+
+        mock_requests.get.assert_called_once_with(
+            file_src.source, stream=True, allow_redirects=True)
+        mock_request.raise_for_status.assert_called_once_with()
+        mock_download.assert_called_once_with(mock_request, file_src.file)
+
+    @unittest.mock.patch(
+        'snapcraft.internal.sources._base.download_urllib_source')
+    def test_download_ftp(self, mock_download):
+        file_src = self.get_mock_file_base(
+            'ftp://snapcraft.io/snapcraft.yaml', 'dir')
+
+        file_src.pull()
+
+        mock_download.assert_called_once_with(file_src.source, file_src.file)
+
+    @unittest.mock.patch('snapcraft.internal.indicators.urlretrieve')
+    def test_download_ftp_url_opener(self, mock_urlretrieve):
+        file_src = self.get_mock_file_base(
+            'ftp://snapcraft.io/snapcraft.yaml', 'dir')
+
+        file_src.pull()
+
+        self.assertEqual(mock_urlretrieve.call_count, 1)
+        self.assertEqual(mock_urlretrieve.call_args[0][0], file_src.source)
+        self.assertEqual(mock_urlretrieve.call_args[0][1], file_src.file)
